==> notebooks/02-General-Preprocessing.ipynb <==
# General Preprocessing


%load_ext autoreload
%autoreload 2

## Change to project directory (Google Colab)

import os 
os.chdir('/content/drive/MyDrive/projects/credit-risk-modelling-udemy/notebooks')
os.listdir()

!pip install -e ../.

## Import Libraries

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np

## Load data

from src.project.config import PROJECT_DIR
import os


fpath = os.path.join(PROJECT_DIR, 'data', 'loan_data_2007_2014_cleaned.csv')
df = pd.read_csv(fpath)
df.head()

## Preprocessing few continuous variables

### emp_length

df['emp_length_int'] = df['emp_length'].str.replace(r'years|year', '').str.strip()
df['emp_length_int'] = df['emp_length_int'].str.replace('< 1', '0')
df['emp_length_int'] = df['emp_length_int'].str.replace(r'10\+', '10')
df['emp_length_int'] = df['emp_length_int'].astype(float).astype('Int64')
df[['emp_length', 'emp_length_int']].drop_duplicates().sort_values('emp_length_int')

* Works as expected :)
* Missing values will be imputed later

### earliest_cr_line

DATE_OF_ANALYSES = pd.to_datetime('2017-12-01', format='%Y-%m-%d')

def convert_to_time_since(ser, ref_date, units, round_val=True):
    time_diff = (pd.to_datetime(ref_date) - pd.to_datetime(ser))/np.timedelta64(1, units)

    if round_val:
        return time_diff.round().astype('Int64')


df['earliest_cr_line'] = pd.to_datetime(df['earliest_cr_line'])
df['mths_since_earliest_cr_line'] = DATE_OF_ANALYSES - df['earliest_cr_line']
df['mths_since_earliest_cr_line'].head()

df['mths_since_earliest_cr_line'] = (df['mths_since_earliest_cr_line']/np.timedelta64(1, 'M')).round().astype('Int64')
df['mths_since_earliest_cr_line'].head()

### term

df['term'].unique()

df['term_int'] = df['term'].str.replace(' months', '').astype(int).astype('Int64')
df['term_int'].head()

### issue_d

df['issue_d'].head()

df['mths_since_issue_d'] = convert_to_time_since(df['issue_d'], DATE_OF_ANALYSES, 'M')
df['mths_since_issue_d'].head()

## Preprocessing few discrete variables

### addr_state

fpath = os.path.join(PROJECT_DIR, 'data/us_states_info.csv')
us_states_info = pd.read_csv(fpath)
us_states_info.head()

us_states_info = pd.merge(us_states_info, df['addr_state'].drop_duplicates(), 
                          left_on='USPS Abbreviation', right_on='addr_state', how='outer', indicator=True)
us_states_info

us_states_info.loc[us_states_info['_merge'] != 'both']

This means North Dakota is missing, and we must add this as a category in the addr_state column just for completeness sake

addr_state_cat = pd.Categorical(df['addr_state'], list(us_states_info['USPS Abbreviation'].dropna()) + ['DC'], ordered=False)
addr_state_cat

df.loc[addr_state_cat.isna(), 'addr_state']

len(addr_state_cat.categories)

pd.get_dummies(addr_state_cat).shape

* SO, yes! pd.get_dummies will add an extra column for the non-existent state

df['addr_state'] = addr_state_cat

## Get Dummy variables for all categorical variables

cat_columns = ['grade', 'sub_grade', 'home_ownership', 'verification_status', 'loan_status',
               'purpose', 'addr_state', 'initial_list_status']

loan_dummies = []
for col in cat_columns:
    dummy_vars = pd.get_dummies(df[col], prefix=col, prefix_sep=':')
    loan_dummies.append(dummy_vars)

loan_dummies = pd.concat(loan_dummies, axis=1)
loan_dummies.head()

### Concat with the original data

df.shape

df = pd.concat([df, loan_dummies], axis=1)
df.shape

## Fill Missing Values

miss_df = df.isna().sum()
miss_df[miss_df != 0]

### total_rev_hi_lim

* Not sure why funded amount is used to fill in the NaNs here...

df['total_rev_hi_lim'].fillna(df['funded_amnt'], inplace=True)

df['annual_inc'].fillna(df['annual_inc'].mean(), inplace=True)

In credit risk, we follow a conservative approach to risk. So we will NaNs with values which would make the loan less suitable (and hence decrease credit score)

df['mths_since_earliest_cr_line'].fillna(0, inplace=True)
df['acc_now_delinq'].fillna(0, inplace=True)
df['total_acc'].fillna(0, inplace=True)
df['pub_rec'].fillna(0, inplace=True)
df['open_acc'].fillna(0, inplace=True)
df['inq_last_6mths'].fillna(0, inplace=True)
df['delinq_2yrs'].fillna(0, inplace=True)
df['emp_length_int'].fillna(0, inplace=True)

miss_df = df.isna().sum()
miss_df[miss_df != 0]

We will ignore the above columns

## loan_status

df['loan_status'].value_counts(normalize=True)

* `Current`, `Fully Paid`, `In Grace Period`, `Does not meet the credit policy. Status:Fully Paid`
    * `Good` Loans - Non-Default
* `Charged Off`, `Late (31-120 days)`, `Late (16-30 days)`, `Default`, `Does not meet the credit policy. Status:Charged Off`
    * `Bad` Loans - Default


good_loan_cats = ['Current', 'Fully Paid', 'In Grace Period', 'Does not meet the credit policy. Status:Fully Paid']
df['good_bad_loan'] = df['loan_status'].isin(good_loan_cats).astype('Int64')
df['good_bad_loan'].value_counts(normalize=True)

* The Dataset is highly imbalanced with only 11% defaults
* Assigning 1 to a good loan is done purposely to ensure that the weights associated with each feature when higher (positive), should imply a positive association with being a good loan
    * and hence in turn imply a higher credit score

out_file = os.path.join(PROJECT_DIR, 'outputs', 'loan_data_general_preprocessed_nb2.pkl')
df.to_pickle(out_file)
