==> notebooks/04-PD-Model-Estimation.ipynb <==
# PD Model Estimation

%load_ext autoreload
%autoreload 2

## Change project directory (google colab)

import os
os.chdir('/content/drive/MyDrive/projects/credit-risk-modelling-udemy/notebooks/')
os.listdir('../')

!pip install -e ../.

## Some imports

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import src
from src.project.config import PROJECT_DIR
from IPython.display import display

## Load Data

fpaths = {subset: os.path.join(PROJECT_DIR, 'outputs', f'loan_data_2007_2014_{subset}_prepared_for_pd_model_nb3.pkl') for subset in ['train', 'test']}
df_train = pd.read_pickle(fpaths['train'])
df_test = pd.read_pickle(fpaths['test'])


df_train.head(2)

df_test.head(2)

## Load Dummy Variable List

infile = os.path.join(PROJECT_DIR, 'outputs', 'dummy_var_lst.csv')
dummy_var_df = pd.read_csv(infile)
dummy_var_df.head()

dummy_var_df['var'] = dummy_var_df['dummy_var'].str.extract(f'(?P<var>.+):.')
dummy_var_df


dummy_var_df.loc[dummy_var_df['var'] == 'mths_since_last_record']

REFERENCE_CATEGORIES = dummy_var_df.drop_duplicates('var').reset_index(drop=True).set_index('var')
REFERENCE_CATEGORIES.loc['mths_since_last_delinq', 'dummy_var'] = 'mths_since_last_delinq:0-4'
REFERENCE_CATEGORIES.loc['mths_since_last_record', 'dummy_var'] = 'mths_since_last_record:0-2'
REFERENCE_CATEGORIES = REFERENCE_CATEGORIES['dummy_var'].tolist()
REFERENCE_CATEGORIES

DUMMY_VARIABLES = pd.Index(dummy_var_df['dummy_var']).difference(REFERENCE_CATEGORIES)
DUMMY_VARIABLES

import json
d = {'DUMMY_VARIABLES': list(DUMMY_VARIABLES), 
     'REFERENCE_CATEGORIES': list(REFERENCE_CATEGORIES)}

out_file = os.path.join(PROJECT_DIR, 'outputs', 'dummy_vars_ref_cats.json')
with open(out_file, 'w') as f:
    json.dump(d, f)

## Prepare $X$ and $y$

X_train, y_train = df_train[DUMMY_VARIABLES], df_train['good_bad_loan'].to_numpy().astype(int)
X_test, y_test = df_test[DUMMY_VARIABLES], df_test['good_bad_loan'].to_numpy().astype(int)

print('Train:', X_train.shape, 'Test:', X_test.shape)

## Probability of Default Modelling using `LogisticRegression`

from sklearn.linear_model import LogisticRegression
import scipy

for varname in LogisticRegression.__init__.__code__.co_varnames:
    print(f'self.{varname} = {varname}')
    


from sklearn.linear_model import LogisticRegression
import numpy as np
import pandas as pd
import scipy

class LogisticRegressionWithPValues(LogisticRegression):

    def fit(self, X, y):
        super().fit(X, y)

        denom = (2.0 * (1.0 + np.cosh(self.decision_function(X))))
        denom = np.tile(denom,(X.shape[1],1)).T
        F_ij = np.dot((X / denom).T,X) ## Fisher Information Matrix
        Cramer_Rao = np.linalg.inv(F_ij) ## Inverse Information Matrix
        sigma_estimates = np.sqrt(np.diagonal(Cramer_Rao))
        z_scores = self.coef_[0] / sigma_estimates # z-score for eaach model coefficient
        p_values = [scipy.stats.norm.sf(abs(x)) * 2 for x in z_scores] ### two tailed test for p-values

        self.p_values = p_values
        try:
            self.feature_names = X.columns
        except AttributeError:
            self.feature_names = [f'X{i}' for i in range(X.shape[1])]

        return self

    def get_summary_table(self, feature_names=None):

        if feature_names is None:
            feature_names = self.feature_names

        # Same as above.
        summary_table = pd.DataFrame(columns = ['feature_name'], data = feature_names)
        summary_table['coefficient'] = np.transpose(self.coef_)
        summary_table.index = summary_table.index + 1
        summary_table.loc[0] = ['intercept', self.intercept_[0]]
        summary_table = summary_table.sort_index()
        summary_table['p_value'] = np.append(np.nan, np.array(self.p_values))
        

        return summary_table

clf = LogisticRegressionWithPValues(solver='saga')
clf.fit(X_train, y_train)

smry_table = clf.get_summary_table()
smry_table

## Write out summary table

out_file = os.path.join(PROJECT_DIR, 'outputs', 'pd_model_logreg_summary_table.csv')
smry_table.to_csv(out_file, index=False)

## Write out Logistic Regression model

import pickle

out_file = os.path.join(PROJECT_DIR, 'outputs', 'pd_model_logreg.pkl')
with open(out_file, 'wb') as f:
    pickle.dump(clf, f)
