==> notebooks/11-Expected-Loss-Calculation.ipynb <==
# Expected Loss Calculation

%load_ext autoreload
%autoreload 2

## Change to project directory (google colab)

import os
os.chdir('/content/drive/MyDrive/projects/credit-risk-modelling-udemy/notebooks')
os.listdir('../')

!pip install -e ../.

## Some imports

import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import src
from src.project.config import PROJECT_DIR

## Load Data

### For LGD and EAD Models

fpath = os.path.join(PROJECT_DIR, 'outputs', 'loan_data_general_preprocessed_nb2.pkl')
loan_data_prep = pd.read_pickle(fpath)
loan_data_prep.head()

loan_data_prep['mths_since_last_delinq'].fillna(0, inplace = True)
loan_data_prep['mths_since_last_record'].fillna(0, inplace = True)


### For PD Model

fpaths = {subset: os.path.join(PROJECT_DIR, 'outputs', f'loan_data_2007_2014_{subset}_prepared_for_pd_model_nb3.pkl') for subset in ['train', 'test']}

loan_data_inputs_train_pd = pd.read_pickle(fpaths['train'])
loan_data_inputs_test_pd = pd.read_pickle(fpaths['test'])
loan_data_inputs_pd = pd.concat([loan_data_inputs_train_pd, loan_data_inputs_test_pd]).reset_index(drop=True)
loan_data_inputs_pd.head()

import json

infile = os.path.join(PROJECT_DIR, 'outputs', 'dummy_vars_ref_cats.json')
with open(infile, 'r') as f:
    d = json.load(f)

DUMMY_VARIABLES_PD = d['DUMMY_VARIABLES']
REFERENCE_CATEGORIES_PD = d['REFERENCE_CATEGORIES']

## Load Models

from src.models import LogisticRegressionWithPValues, LinearRegressionWithPValues
import pickle

### PD Model

fpath = os.path.join(PROJECT_DIR, 'outputs', 'pd_model_logreg.pkl')
with open(fpath, 'rb') as f:
    pd_model = pickle.load(f)


### LGD Model - Stage 1

fpath = os.path.join(PROJECT_DIR, 'outputs', 'lgd_model_stage_1.pkl')
with open(fpath, 'rb') as f:
    reg_lgd_stage_1 = pickle.load(f)


### LGD Model - Stage 2

fpath = os.path.join(PROJECT_DIR, 'outputs', 'lgd_model_stage_2.pkl')
with open(fpath, 'rb') as f:
    reg_lgd_stage_2 = pickle.load(f)


### EAD Model

fpath = os.path.join(PROJECT_DIR, 'outputs', 'ead_model.pkl')
with open(fpath, 'rb') as f:
    reg_ead = pickle.load(f)


## Prepare Inputs for `LGD` an `EAD` models

## Variable List

variable_lst = ['grade:A',
'grade:B',
'grade:C',
'grade:D',
'grade:E',
'grade:F',
'grade:G',
'home_ownership:MORTGAGE',
'home_ownership:NONE',
'home_ownership:OTHER',
'home_ownership:OWN',
'home_ownership:RENT',
'verification_status:Not Verified',
'verification_status:Source Verified',
'verification_status:Verified',
'purpose:car',
'purpose:credit_card',
'purpose:debt_consolidation',
'purpose:educational',
'purpose:home_improvement',
'purpose:house',
'purpose:major_purchase',
'purpose:medical',
'purpose:moving',
'purpose:other',
'purpose:renewable_energy',
'purpose:small_business',
'purpose:vacation',
'purpose:wedding',
'initial_list_status:f',
'initial_list_status:w',
'term_int',
'emp_length_int',
'mths_since_issue_d',
'mths_since_earliest_cr_line',
'funded_amnt',
'int_rate',
'installment',
'annual_inc',
'dti',
'delinq_2yrs',
'inq_last_6mths',
'mths_since_last_delinq',
'mths_since_last_record',
'open_acc',
'pub_rec',
'total_acc',
'acc_now_delinq',
'total_rev_hi_lim']


variable_ref_cat =  ['grade:G',
'home_ownership:RENT',
'verification_status:Verified',
'purpose:credit_card',
'initial_list_status:f']

loan_data_prep_lgd_ead = loan_data_prep[variable_lst].drop(variable_ref_cat, axis=1)

### Recovery Rate


loan_data_prep['recovery_rate_stage_1'] = reg_lgd_stage_1.predict(loan_data_prep_lgd_ead)  # Stage 1: predict whether recovery_rate 0 or not
loan_data_prep['recovery_rate_stage_2'] = reg_lgd_stage_2.predict(loan_data_prep_lgd_ead) # Stage 2: predict the actual recovery rate given non-zero
loan_data_prep['recovery_rate'] = loan_data_prep['recovery_rate_stage_1'] * loan_data_prep['recovery_rate_stage_2'] # Combining Stages

loan_data_prep['recovery_rate'] = np.where(loan_data_prep['recovery_rate'] < 0, 0, loan_data_prep['recovery_rate'])
loan_data_prep['recovery_rate'] = np.where(loan_data_prep['recovery_rate'] > 1, 1, loan_data_prep['recovery_rate'])

### LGD

loan_data_prep['lgd'] = 1 - loan_data_prep['recovery_rate']

### CCF

loan_data_prep['ccf'] = reg_ead.predict(loan_data_prep_lgd_ead)

loan_data_prep['ccf'] = np.where(loan_data_prep['ccf'] < 0, 0, loan_data_prep['ccf'])
loan_data_prep['ccf'] = np.where(loan_data_prep['ccf'] > 1, 1, loan_data_prep['ccf'])

### EAD

loan_data_prep['ead'] = loan_data_prep['ccf']*loan_data_prep['funded_amnt']

### PD

loan_data_inputs_pd['pd'] = pd_model.predict_proba(loan_data_inputs_pd[DUMMY_VARIABLES_PD])[:, 0]

## Combining the features of the two dataframes

loan_data_prep_comb = pd.concat([loan_data_prep.set_index('index'), loan_data_inputs_pd.set_index('index')], axis=1)

loan_data_prep_comb.shape

### EL

loan_data_prep_comb['el'] = loan_data_prep_comb['pd']*loan_data_prep_comb['lgd']*loan_data_prep_comb['ead']

loan_data_prep_comb[['funded_amnt', 'pd', 'lgd', 'ead', 'el']].head().T.drop_duplicates().T

loan_data_prep_comb['el'].sum()

loan_data_prep_comb['el'].sum() / loan_data_prep_comb['funded_amnt'].sum().drop_duplicates()


* 7.67% is the Expected Loss on the loan portfolio which is a bit less than what a bank holds in its capital, which is around 10%. 
* Observed Expected Loss values are anywhere between 2% and 10%
* Depending upon the exposure, the bank management can decide to be conservative or aggressive when giving out loans
